Fix the inappropriate Play Services update request
The issue is with '+' dependency references. The Fit Preview Play
Services (v5.2) gets picked by Gradle as it's available on the user's
local repo, but it's not available on devices (except for devices with
the L Fit Preview). Devices out there only have 5.0.89 available.

Swapping out all '+' dependencies, the app now correctly works and
builds no matter what the environment is. This is also recommended by
the new Google guidelines.