Map topic photos to Track tags.
Each track tag requires artwork, the artwork is defined on all
topics that are in that track. This change extracts the topic
artwork from the first topic that contains the Track tag as
a value in its array of categoryIds. A topic cannot be associated
with more than one Track.

Change-Id: Ied02dc16f311e199d70fcbdab31edc714c08e104
Bug: 28143531