Fixing test build.
Many of the tests marked suppressed failed when run
in an instrumentation with all tests. Those tests
succeed when run only as part of the test package.
I've disabled them anyway because the test build need
to pass with all tests run together.

The test library updates are required to pickup some bug fixes
that were causing some of the test issues.

Bug: 28369715
Change-Id: Ic1abd17dc5ae2102ad8e487cbca4db2029ea65e2